fix: move test classes to individual files
